requirements-cpu.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
attrs==24.2.0
audioread==3.0.1
audiostretchy==1.3.5
babel==2.16.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comtypes==1.4.8
confection==0.1.5
contourpy==1.3.1
coqpit==0.0.17
coqui-tts==0.24.3
coqui-tts-trainer==0.1.6
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
dxcam==0.0.5
einops==0.8.0
elevenlabs==1.12.1
encodec==0.1.1
filelock==3.13.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.2.0
greenlet==3.1.1
grpcio==1.67.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
inflect==7.4.0
Jinja2==3.1.3
jiter==0.7.1
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-core==0.3.17
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langcodes==3.4.1
langsmith==0.1.142
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
networkx==3.2.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
openai==1.54.4
opencv-python==4.10.0.84
orjson==3.10.11
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pypiwin32==223
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
SudachiDict-core==20241021
SudachiPy==0.6.8
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
tqdm==4.67.0
transformers==4.46.2
typeguard==4.4.1
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.17.1
